requirements.txt
Flask==1.1.2
opencensus-ext-azure==1.0.4
opencensus-ext-flask==0.7.3
redis==3.5.3
opencensus==0.7.13
opencensus-ext-logging==0.1.0
azure.identity
azure.mgmt.compute
